Phase 2: Remove I.E + Hide 1-5 rating + Change Y/N to R.A.K/L.A.K
Try make dropdown look better. Remove name “Information explorer” - not needed. + Don’t need to show the 1-5 rating for them to input as it interferes with brain signal as they read more + Change y/n to left or right arrow keys to make it easier for someone to respond naturally rather than having to check for the letter keys on the computer.